split files out to two modules (consumer/producer) and a main
